moved tests to init
